+++ test/object-store/migrations.cpp
@@ -926,6 +926,64 @@ TEST_CASE("migration: Automatic") {
             }
         }
 
+        SECTION("change table to embedded - multiple incoming links per object resolved by removing a column") {
+            Schema schema = {
+                {"child_table",
+                 {
+                     {"value", PropertyType::Int},
+                 }},
+                {"parent_table",
+                 {
+                     {"child_property", PropertyType::Object | PropertyType::Nullable, "child_table"},
+                     {"child_property_duplicate", PropertyType::Object | PropertyType::Nullable, "child_table"},
+                 }},
+            };
+            Schema schema2 = {
+                {"child_table",
+                 IsEmbedded{true},
+                 {
+                     {"value", PropertyType::Int},
+                 }},
+                {"parent_table",
+                 {
+                     {"child_property", PropertyType::Object | PropertyType::Nullable, "child_table"},
+                 }},
+            };
+
+            auto realm = Realm::get_shared_realm(config);
+            realm->update_schema(schema, 1);
+            realm->begin_transaction();
+            auto child_table = ObjectStore::table_for_object_type(realm->read_group(), "child_table");
+            Obj child_object1 = child_table->create_object();
+            child_object1.set("value", 42);
+            Obj child_object2 = child_table->create_object();
+            child_object2.set("value", 43);
+            auto parent_table = ObjectStore::table_for_object_type(realm->read_group(), "parent_table");
+            auto child_object_key1 = child_object1.get_key();
+            auto child_object_key2 = child_object2.get_key();
+            parent_table->create_object().set_all(child_object_key1, child_object_key1);
+            parent_table->create_object().set_all(child_object_key2, child_object_key2);
+            realm->commit_transaction();
+            REQUIRE(parent_table->size() == 2);
+            REQUIRE(child_table->size() == 2);
+            REQUIRE_FALSE(child_table->is_embedded());
+
+            REQUIRE_NOTHROW(realm->update_schema(schema2, 2, nullptr));
+
+            REQUIRE(realm->schema_version() == 2);
+            REQUIRE(parent_table->size() == 2);
+            REQUIRE(child_table->size() == 2);
+            REQUIRE(child_table->is_embedded());
+            CppContext context(realm);
+            for (int i = 0; i < 2; i++) {
+                Object parent_object(realm, "parent_table", i);
+                Object child_object =
+                    any_cast<Object>(parent_object.get_property_value<util::Any>(context, "child_property"));
+                Int value = any_cast<Int>(child_object.get_property_value<util::Any>(context, "value"));
+                REQUIRE(value == 42 + i);
+            }
+        }
+
         SECTION("change table to embedded - multiple incoming links - resolved in migration block") {
             Schema schema = {
                 {"child_table",